pyproject.toml
[tool.poetry]
name = "github-kb"
version = "0.1.0"
description = "Ask questions about a GitHub repository with Strands Agents and AWS Bedrock"
authors = ["Gonzalo Ayuso <gonzalo123@gmail.com>"]
readme = "README.md"
packages = [{include = "github_kb", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
boto3 = "^1.42.78"
click = "^8.3.1"
pydantic = "^2.12.3"
pydantic-settings = "^2.13.1"
python-dotenv = "^1.2.2"
rich = "^14.3.0"
strands-agents = "^1.34.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"

[tool.poetry.scripts]
github-kb = "github_kb.cli:cli"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
